+++ src/models/oauth/queries.rs
@@ -40,6 +40,56 @@ pub async fn create_oauth_app(
     Ok(app)
 }
 
+pub async fn get_oauth_app_by_client_id(
+    db_client: &impl DatabaseClient,
+    client_id: &Uuid,
+) -> Result<DbOauthApp, DatabaseError> {
+    let maybe_row = db_client.query_opt(
+        "
+        SELECT oauth_application
+        FROM oauth_application
+        WHERE client_id = $1
+        ",
+        &[&client_id],
+    ).await?;
+    let row = maybe_row.ok_or(DatabaseError::NotFound("oauth application"))?;
+    let app = row.try_get("oauth_application")?;
+    Ok(app)
+}
+
+pub async fn create_oauth_authorization(
+    db_client: &impl DatabaseClient,
+    authorization_code: &str,
+    user_id: &Uuid,
+    application_id: i32,
+    scopes: &str,
+    created_at: &DateTime<Utc>,
+    expires_at: &DateTime<Utc>,
+) -> Result<(), DatabaseError> {
+    db_client.execute(
+        "
+        INSERT INTO oauth_authorization (
+            code,
+            user_id,
+            application_id,
+            scopes,
+            created_at,
+            expires_at
+        )
+        VALUES ($1, $2, $3, $4, $5, $6)
+        ",
+        &[
+            &authorization_code,
+            &user_id,
+            &application_id,
+            &scopes,
+            &created_at,
+            &expires_at,
+        ],
+    ).await?;
+    Ok(())
+}
+
 pub async fn save_oauth_token(
     db_client: &impl DatabaseClient,
     owner_id: &Uuid,
@@ -139,6 +189,31 @@ mod tests {
         };
         let app = create_oauth_app(db_client, db_app_data).await.unwrap();
         assert_eq!(app.app_name, "My App");
+    }
+
+    #[tokio::test]
+    #[serial]
+    async fn test_create_oauth_authorization() {
+        let db_client = &mut create_test_database().await;
+        let user_data = UserCreateData {
+            username: "test".to_string(),
+            ..Default::default()
+        };
+        let user = create_user(db_client, user_data).await.unwrap();
+        let app_data = DbOauthAppData {
+            app_name: "My App".to_string(),
+            ..Default::default()
+        };
+        let app = create_oauth_app(db_client, app_data).await.unwrap();
+        create_oauth_authorization(
+            db_client,
+            "code",
+            &user.id,
+            app.id,
+            "read write",
+            &Utc::now(),
+            &Utc::now(),
+        ).await.unwrap();
     }
 
     #[tokio::test]